chore(photos): consolidation pass — AnnIndex::is_empty + dead-cast cleanup (Track C)
Clippy on the photos module surfaces no correctness/robustness findings (only
style lints) — a good stability signal. Genuine tidy-ups: completed the
AnnIndex API with is_empty(); removed two unnecessary same-type casts
(hjson::now_iso civil-date, stitch cross-fade). Docs (KEYMAP) confirmed current
for the 3.10-3.13 additions. 145 photos tests.

+++ src/photos/stitch.rs
@@ -213,7 +213,7 @@ fn stitch_aligned_h(parts: &[RgbImage]) -> RgbImage {
                     continue;
                 }
                 let src = b.get_pixel(bx, by).0;
-                let px = if (bx as u32) < overlap && ox < cw as i64 {
+                let px = if bx < overlap && ox < cw as i64 {
                     // Linear cross-fade: 0 at the seam start (keep canvas) → 1 at overlap end (take b).
                     let t = (bx as f32 + 1.0) / (overlap as f32 + 1.0);
                     let dst = out.get_pixel(ox as u32, oy as u32).0;
